完成支付界面

+++ 后端/src/main/java/com/turni/lifepayapp/bean/Bill2.java
@@ -14,6 +14,6 @@ public class Bill2 {
     private Integer day;
     private Integer hour;
     private Integer minute;
-    private Integer money;
+    private Double money;
     private String name;
 }
